# Get faster response for count() method